+++ .circleci/images/Dockerfile
@@ -1,4 +1,4 @@
-FROM ubuntu:18.04
+FROM ubuntu:19.10
 MAINTAINER rfletcher@esri.com
 RUN apt-get update && apt-get -y install git ssh tar gzip ca-certificates gdal-bin gdal-data libgdal-dev pkg-config
 RUN wget https://dl.google.com/go/go1.14.1.linux-amd64.tar.gz && tar -xzvf go1.14.1.linux-amd64.tar.gz && mv go /usr/local
