+++ macro/mosaic_test.go
@@ -1,15 +1,15 @@
 package macro
 
 import (
-	"github.com/boundlessgeo/go-gdal"
 	"testing"
+
+	"github.com/Rob-Fletcher/go-gdal"
 )
 
-
 const S3_BUCKET_URL = "/vsis3/imagerytest"
-const S3_FILE  = "/vsis3/imagerytest/m_3707501_ne_18_1_20150801.tif"
+const S3_FILE = "/vsis3/imagerytest/m_3707501_ne_18_1_20150801.tif"
 
-func TestMosaic(t *testing.T) {
+func testMosaic(t *testing.T) {
 	type args struct {
 		urls []string
 		epsg string
@@ -18,7 +18,7 @@ func TestMosaic(t *testing.T) {
 		name string
 		args args
 	}{
-		{ name: "everything", args: args{urls: gdal.VSIReadDirRecursiveAbs(S3_BUCKET_URL), epsg: "EPSG:4326"}},
+		{name: "everything", args: args{urls: gdal.VSIReadDirRecursiveAbs(S3_BUCKET_URL), epsg: "EPSG:4326"}},
 	}
 
 	for _, tt := range tests {
